Moved logic to its respective routes and also fixed a bug with the system edit route
