+++ src/routing/routes/AdsorbateDetailRoute.jsx
@@ -1,16 +1,18 @@
-import {useDispatch, useSelector} from "react-redux";
 import React, {useEffect} from "react";
+import {useParams} from "react-router-dom";
+import {useDispatch, useSelector} from "react-redux";
 import CircularProgress from "@material-ui/core/CircularProgress";
 import {fetchAdsorbate} from "../../redux/adsorbateSlice";
 import {AdsorbateDetail} from "../../components/Detail/AdsorbateDetail";
 import {Redirect} from "react-router-dom";
-import {URLS} from "../../routing/urls";
+import {URLS} from "../urls";
 import {errorCodes} from "../../utils/errorStatusCodes";
 
-export const AdsorbateDetailContainer = ({adsorbateId}) => {
+export const AdsorbateDetailRoute = () => {
+  let {id} = useParams();
   const dispatch = useDispatch();
   useEffect(() => {
-    dispatch(fetchAdsorbate(adsorbateId));
+    dispatch(fetchAdsorbate(id));
   }, []);
 
   const adsorbate = useSelector((store) => store.adsorbate.adsorbate);
