--- src/routing/routes/AdsorbatesRoute.js
@@ -1,6 +1,0 @@
-import React from "react";
-import {AdsorbateListContainer} from "../../containers/List/AdsorbateListContainer";
-
-export const AdsorbatesRoute = () => {
-  return <AdsorbateListContainer />;
-};
