--- src/routing/routes/AdsorbentCreateRoute.js
@@ -1,6 +1,0 @@
-import React from "react";
-import {AdsorbentCreateContainer} from "../../containers/Create/AdsorbentCreateContainer";
-
-export const AdsorbentCreateRoute = () => {
-  return <AdsorbentCreateContainer />;
-};
