+++ src/routing/routes/AdsorbentEditRoute.jsx
@@ -1,9 +1,63 @@
 import {useParams} from "react-router-dom";
-import React from "react";
-import {AdsorbentEditForm} from "../../components/CustomForms/AdsorbentEditForm";
+import React, {useEffect, useState} from "react";
+import {useDispatch, useSelector} from "react-redux";
+import {fetchAdsorbent} from "../../redux/adsorbentSlice";
+import {editAdsorbent} from "../../services/adsorbents";
+import {URLS} from "../../routing/urls";
+import {useHistory} from "react-router-dom";
+import {AdsorbentForm} from "../../components/CustomForms/AdsorbentForm";
+import {Redirect} from "react-router-dom";
+import {errorCodes} from "../../utils/errorStatusCodes";
+import {displayUpdateMessage} from "../../utils/displayUpdateMessage";
+import {FormErrorModal} from "../../components/Form/FormErrorModal";
 
 export const AdsorbentEditRoute = () => {
   let {id} = useParams();
+  const [errors, setErrors] = useState(true);
+  const [submitError, setSubmitError] = useState(null);
 
-  return <AdsorbentEditForm id={id} />;
+  const dispatch = useDispatch();
+  const history = useHistory();
+
+  const adsorbent = useSelector((store) => store.adsorbent.adsorbent);
+  useEffect(() => {
+    if (!adsorbent) {
+      dispatch(fetchAdsorbent(id));
+    }
+  }, []);
+
+  const onSubmit = async (values) => {
+    if (!errors) {
+      try {
+        const result = await editAdsorbent(id, {...values, id: parseInt(id)});
+        if (result.status) {
+          setSubmitError(result.response);
+        } else {
+          displayUpdateMessage();
+          history.push(URLS.ADSORBENTS_LIST);
+        }
+      } catch (error) {
+        return error;
+      }
+    }
+  };
+
+  if (adsorbent && errorCodes.includes(adsorbent.status)) {
+    return <Redirect to={URLS.NOT_FOUND} />;
+  }
+  return (
+    <>
+      <FormErrorModal
+        errorInfo={submitError && submitError.message}
+        onClose={() => setSubmitError(null)}
+      />
+      <AdsorbentForm
+        title={"Modificar adsorbente"}
+        onSubmit={onSubmit}
+        buttonLabel={"Actualizar"}
+        setErrors={setErrors}
+        initialValues={adsorbent}
+      />
+    </>
+  );
 };
