--- src/routing/routes/AdsorbentsRoute.js
@@ -1,6 +1,0 @@
-import React from "react";
-import {AdsorbentListContainer} from "../../containers/List/AdsorbentListContainer";
-
-export const AdsorbentsRoute = () => {
-  return <AdsorbentListContainer />;
-};
