+++ src/routing/routes/AdsorbentsRoute.jsx
@@ -1,13 +1,16 @@
 import React, {useEffect, useState} from "react";
-import {AdsorbentSearchContainer} from "./Search/AdsorbentSearchContainer";
+import {AdsorbentSearchContainer} from "../../containers/List/Search/AdsorbentSearchContainer";
 import {AdsorbentList} from "../../components/List/AdsorbentList/AdsorbentList";
 import {ListHeader} from "../../components/List/common/ListHeader";
 import {URLS} from "../../routing/urls";
 import {getAdsorbents} from "../../services/adsorbents";
+import {useSelector} from "react-redux";
 
-export function AdsorbentListContainer() {
+export const AdsorbentsRoute = () => {
   const [loading, setLoading] = useState(false);
   const [adsorbents, setAdsorbents] = useState(null);
+  const storedAdsorbents = useSelector((state) => state.adsorbents.adsorbents);
+
   useEffect(async () => {
     setLoading(true);
     const response = await getAdsorbents();
@@ -17,6 +20,10 @@ export function AdsorbentListContainer() {
     setLoading(false);
   }, []);
 
+  useEffect(async () => {
+    setAdsorbents(storedAdsorbents);
+  }, [storedAdsorbents]);
+
   return (
     <>
       <ListHeader title={"Adsorbentes"} creationUrl={URLS.ADSORBENT_CREATE} />
@@ -24,4 +31,4 @@ export function AdsorbentListContainer() {
       <AdsorbentList loading={loading} adsorbents={adsorbents} />
     </>
   );
-}
+};
