--- src/routing/routes/CalculateVolumeRoute.js
@@ -1,6 +1,0 @@
-import React from "react";
-import {CalculateVolumeViewContainer} from "../../containers/CalculateVolumeViewContainer";
-
-export const CalculateVolumeRoute = () => {
-  return <CalculateVolumeViewContainer />;
-};
