--- src/routing/routes/IdealAdsorbentRoute.js
@@ -1,6 +1,0 @@
-import React from "react";
-import {IdealAdsorbentContainer} from "../../containers/List/IdealAdsorbentContainer";
-
-export const IdealAdsorbentRoute = () => {
-  return <IdealAdsorbentContainer />;
-};
