+++ src/routing/routes/IdealAdsorbentRoute.jsx
@@ -1,14 +1,17 @@
 import React, {useEffect} from "react";
 import Container from "@material-ui/core/Container";
-import {IdealAdsorbentSearchContainer} from "./Search/IdealAdsorbentSearchContainer";
+import {IdealAdsorbentSearchContainer} from "../../containers/List/Search/IdealAdsorbentSearchContainer";
 import {fetchAdsorbatesWithIupacNotation} from "../../redux/adsorbatesSlice";
 import {useDispatch, useSelector} from "react-redux";
 import Typography from "@material-ui/core/Typography";
 import {IdealAdsorbentList} from "../../components/IdealAdsorbentList/IdealAdsorbentList";
-import {TittleContainer, IdealAdsorbentListContainer} from "./Styles";
+import {
+  TittleContainer,
+  IdealAdsorbentListContainer,
+} from "../../containers/List/Styles";
 import {resetIdealAdsorbents} from "../../redux/idealAdsorbentsSlice";
 
-export function IdealAdsorbentContainer() {
+export const IdealAdsorbentRoute = () => {
   const loading = useSelector((state) => state.loading);
   const {idealAdsorbents} = useSelector((state) => state.idealAdsorbents);
 
@@ -34,4 +37,4 @@ export function IdealAdsorbentContainer() {
       </IdealAdsorbentListContainer>
     </Container>
   );
-}
+};
