--- src/routing/routes/ProcessesRoute.js
@@ -1,6 +1,0 @@
-import React from "react";
-import {ProcessListContainer} from "../../containers/List/ProcessListContainer";
-
-export const ProcessesRoute = () => {
-  return <ProcessListContainer />;
-};
